test: split test_slurm.py into tests/tools/slurm/{test_parse,test_errors,test_resources}.py
